Fix a regression.
